+++ pmm-app/src/pmm-qan/panel/components/Filters/Filters.tsx
@@ -1,6 +1,4 @@
-import React, {
-  FC, useContext, useEffect, useMemo, useRef, useState,
-} from 'react';
+import React, { FC, useContext, useEffect, useMemo, useRef, useState } from 'react';
 import { Button, Input, Spin } from 'antd';
 import { Form } from 'react-final-form';
 import { cx } from 'emotion';
@@ -10,10 +8,12 @@ import { Scrollbar } from 'shared/components/Elements/Scrollbar/Scrollbar';
 import { useTheme } from '@grafana/ui';
 import { CheckboxGroup } from './components/CheckboxGroup/CheckboxGroup';
 import { FILTERS_BODY_HEIGHT, FILTERS_GROUPS } from './Filters.constants';
-import { useFilters, useFiltersContainerHeight, useInitialFilterValues } from './Filters.hooks';
 import { getSelectedCheckboxes } from './Filters.tools';
 import { FiltersContainerProps } from './Filters.types';
 import { getStyles } from './Filters.styles';
+import { useFilters } from './hooks/useFilters';
+import { useInitialFilterValues } from './hooks/useInitialFilterValues';
+import { useFiltersContainerHeight } from './hooks/useFiltersContainerHeight';
 
 export const FiltersContainer: FC<FiltersContainerProps> = ({
   contextActions,
@@ -38,46 +38,69 @@ export const FiltersContainer: FC<FiltersContainerProps> = ({
     }
   }, [selectedCheckboxes]);
 
-  return (
-    <div ref={filtersWrapperRef} className={cx({ [styles.filtersDisabled]: disabled })}>
+  const ShowAllButton = () => {
+    return (
+      <Button
+        type="link"
+        className={styles.showAllButton}
+        onClick={() => showSetAll(!showAll)}
+        disabled={!selectedCheckboxes}
+        data-qa="qan-filters-show-selected"
+      >
+        {showAll ? 'Show Selected' : 'Show All'}
+      </Button>
+    );
+  };
+
+  const ResetButton = () => {
+    return (
+      <Button
+        type="link"
+        htmlType={'reset'}
+        className={styles.resetButton}
+        data-qa="qan-filters-reset-all"
+        onClick={() => {
+          setFilter('');
+          showSetAll(true);
+        }}
+        disabled={!selectedCheckboxes}
+      >
+        Reset All
+      </Button>
+    );
+  };
+
+  const FiltersHeader = () => {
+    return (
       <div className={styles.filtersHeader}>
         <h5 className={styles.title}>Filters</h5>
-        <Button
-          type="link"
-          className={styles.showAllButton}
-          onClick={() => showSetAll(!showAll)}
-          disabled={!selectedCheckboxes}
-          data-qa="qan-filters-show-selected"
-        >
-          {showAll ? 'Show Selected' : 'Show All'}
-        </Button>
-        <Button
-          type="link"
-          className={styles.resetButton}
-          data-qa="qan-filters-reset-all"
-          onClick={() => {
-            setFilter('');
-            showSetAll(true);
-            contextActions.resetLabels();
-            form.reset();
-          }}
-          disabled={!selectedCheckboxes}
-        >
-          Reset All
-        </Button>
+        <ShowAllButton />
+        <ResetButton />
       </div>
+    );
+  };
+
+  const FilterInput = () => {
+    return (
+      <Input
+        suffix={<Filter fill="#c6c6c6" />}
+        placeholder="Filter by..."
+        onChange={(e) => {
+          setFilter(e.target.value);
+          e.stopPropagation();
+        }}
+        value={filter}
+        className={styles.filtersField}
+        data-qa="filters-search-field"
+      />
+    );
+  };
+
+  return (
+    <div ref={filtersWrapperRef} className={cx({ [styles.filtersDisabled]: disabled })}>
+      <FiltersHeader />
       <Scrollbar className={styles.getFiltersWrapper(height)}>
-        <Input
-          suffix={<Filter fill="#c6c6c6" />}
-          placeholder="Filter by..."
-          onChange={(e) => {
-            setFilter(e.target.value);
-            e.stopPropagation();
-          }}
-          value={filter}
-          className={styles.filtersField}
-          data-qa="filters-search-field"
-        />
+        <FilterInput />
         {FILTERS_GROUPS.filter((group) => filters[group.dataKey]).map(
           ({ name, dataKey, getDashboardURL }) => (
             <CheckboxGroup
@@ -112,9 +135,16 @@ export const Filters: FC = () => {
         initialValues={initialValues}
         render={({ form, handleSubmit }) => (
           <Spin spinning={loading}>
-            <form onSubmit={handleSubmit} onChange={() => contextActions.setLabels(form.getState().values)}>
+            <form
+              onSubmit={handleSubmit}
+              onChange={() => {
+                contextActions.setLabels(form.getState().values);
+              }}
+              onReset={() => {
+                contextActions.resetLabels();
+              }}
+            >
               <FiltersContainer
-                contextActions={contextActions}
                 form={form}
                 filters={filters}
                 disabled={loadingDetails}
