+++ pmm-app/src/pmm-qan/panel/components/Filters/Filters.types.ts
@@ -1,7 +1,6 @@
 import { RawTimeRange } from '@grafana/data';
 
 export interface FiltersContainerProps {
-  contextActions: any;
   form: any;
   filters: any;
   disabled?: boolean;
